+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file. The format 
 
 - Fixed a bug that made pressing Enter on the access request page clear the form
   instead of submitting it.
+- Fixed anchor link positioning.
 
 ## [3.6.0] 2023-05-05